Refactor WS812FX file names
